+++ task/build.lua
@@ -14,6 +14,7 @@ map: `settings.map.input` and `settings.map.output` must differ]]
 		end
 
 		table.insert (tasks, 'build.process-imports')
+		table.insert (tasks, 'environment.teardown')
 		table.insert (tasks, 'build.w3x.write')
 	end