+++ task/check.lua
@@ -2,12 +2,8 @@ local Tasks = require ('map.tasks')
 
 return function (state)
 	local tasks = {
-		'build.environment'
+		'environment.setup'
 	}
-
-	if state.settings.map then
-		table.insert (tasks, 'build.w3x.read')
-	end
 
 	if state.settings.build then
 		table.insert (tasks, 'build.user-files')